feat: implement global KV cache allocation for Qwen3
Implements a global KV cache allocation system for efficient memory management in the inference server.

Changes:
- Add global KV cache allocator that pre-allocates contiguous GPU memory for all attention layers
- Add --block-size and --memory-utilization CLI arguments for cache configuration
- Add prompt_token_ids and generated_token_ids fields to Sequence for token tracking
- Update ModelConfig with block_size and memory_utilization parameters
- Update dtype defaults to bfloat16 and use dtype parameter instead of torch_dtype

+++ server/model/hf_runner.py
@@ -283,7 +283,7 @@ def load_hf_model(config: ModelConfig) -> ModelRunner:
     )
     model: PreTrainedModel = AutoModelForCausalLM.from_pretrained(
         config.model_name_or_path,
-        torch_dtype=config.dtype,
+        dtype=config.dtype,
         device_map="auto" if config.device == "cuda" else None,
     )
